fix: refactor input validation to its own function
