+++ src/constants.ts
@@ -1,0 +1,11 @@
+// SPDX-FileCopyrightText: 2022 - 2024 Ali Sajid Imami
+//
+// SPDX-License-Identifier: MIT
+
+// A file to store and manage the constants
+
+// Set the minimum time limit to 0 seconds
+export const MINIMUM_ALLOWED: number = 0
+
+// Set the maximum time limit to 2 minutes (120 seconds)
+export const MAXIMUM_ALLOWED: number = 120
